Remove InputParameter.java Remove Parser Interface Rename CliArgumentParser to CliArgument Make CliArgument an object representation of the user-supplied arguments
