+++ src/main/java/reposense/parser/CliArguments.java
@@ -1,5 +1,6 @@
 package reposense.parser;
 
+import java.io.File;
 import java.util.Date;
 import java.util.HashMap;
 import java.util.Optional;
@@ -7,9 +8,9 @@ import java.util.Optional;
 import reposense.exception.ParseException;
 
 /**
- * Verifies and parses user-supplied CLI arguments into an InputParameter Object.
+ * Verifies and parses user-supplied CLI arguments.
  */
-public class CliArgumentsParser implements Parser<InputParameter, String[]> {
+public class CliArguments {
 
     private static final String ARGUMENT_PREFIX_DASH = "-";
 
@@ -21,15 +22,24 @@ public class CliArgumentsParser implements Parser<InputParameter, String[]> {
     private static final String PARSE_EXCEPTION_MESSAGE_INVALID_INPUTS = "Failed to parse inputs arguments";
     private static final String PARSE_EXCEPTION_MESSAGE_NO_CSV_FILE = "Failed due to missing CSV file";
 
+    private static final String DEFAULT_FILE_ARG = ".";
+
+    private File configFile;
+    private File targetFile;
+    private Optional<Date> sinceDate;
+    private Optional<Date> untilDate;
+
     /**
-     * Parses user-supplied arguments into an InputParameter object.
+     * Parses user-supplied arguments.
      *
      * @throws ParseException If the given inputs or the csv file fail to parse or mandatory fields are missing.
      */
-    @Override
-    public InputParameter parse(String[] args) throws ParseException {
+    public CliArguments(String[] args) throws ParseException {
+        sinceDate = Optional.empty();
+        untilDate = Optional.empty();
+        targetFile = new File(DEFAULT_FILE_ARG);
+
         final HashMap<String, String> argumentMap = new HashMap<String, String>();
-        final InputParameter parameter = new InputParameter();
 
         for (int i = 0; i < args.length; i = i + 2) {
 
@@ -46,10 +56,10 @@ public class CliArgumentsParser implements Parser<InputParameter, String[]> {
         }
 
         checkAllMandatoryArgumentsPresent(argumentMap);
-        setUserInputValuesToArgument(argumentMap, parameter);
-
-        return parameter;
+        setUserInputValuesToArgument(argumentMap);
     }
+
+
 
     /**
      * Checks if all mandatory fields are present.
@@ -63,34 +73,49 @@ public class CliArgumentsParser implements Parser<InputParameter, String[]> {
     }
 
     /**
-     * Set user-supplied inputs to the InputParameter object
+     * Set user-supplied inputs to instance variables
      *
      * @throws ParseException If the supplied dates fail to parse.
      */
-    private void setUserInputValuesToArgument(final HashMap<String, String> argumentMap, final InputParameter parameter)
+    private void setUserInputValuesToArgument(final HashMap<String, String> argumentMap)
             throws ParseException {
         final DateParser dateParser = new DateParser();
 
         if (argumentMap.containsKey(CONFIG_FILE_ARG)) {
-            final String configFile = argumentMap.get(CONFIG_FILE_ARG);
-            parameter.setConfigFile(configFile);
+            final String configFilePath = argumentMap.get(CONFIG_FILE_ARG);
+            configFile = new File(configFilePath);
         }
 
         if (argumentMap.containsKey(TARGET_FILE_ARG)) {
-            final String targetFile = argumentMap.get(TARGET_FILE_ARG);
-            parameter.setTargetFile(targetFile);
+            final String targetFilePath = argumentMap.get(TARGET_FILE_ARG);
+            targetFile = new File(targetFilePath);
         }
 
         if (argumentMap.containsKey(SINCE_DATE_ARG)) {
             final String sinceDateInString = argumentMap.get(SINCE_DATE_ARG);
-            final Optional<Date> sinceDate = dateParser.parse(sinceDateInString);
-            parameter.setSinceDate(sinceDate.orElse(null));
+            sinceDate = dateParser.parse(sinceDateInString);
         }
 
         if (argumentMap.containsKey(UNTIL_DATE_ARG)) {
             final String untilDateInString = argumentMap.get(UNTIL_DATE_ARG);
-            final Optional<Date> untilDate = dateParser.parse(untilDateInString);
-            parameter.setUntilDate(untilDate.orElse(null));
+            untilDate = dateParser.parse(untilDateInString);
         }
     }
+
+    public File getConfigFile() {
+        return configFile;
+    }
+
+    public Optional<Date> getSinceDate() {
+        return sinceDate;
+    }
+
+    public File getTargetFile() {
+        return targetFile;
+    }
+
+    public Optional<Date> getUntilDate() {
+        return untilDate;
+    }
+
 }
